Remove unused 'border' squares in board
Border squares were intended to make it easier/faster to tell when a
piece could not move in a given direction. Not using this idea, so
reduced the board to just 8 * 8 squares. This will hopefully make it
faster to hash a position, should we store a table of position to score
info.