First stab at iterative deepening
Search function now performs a depth-first search to depth 1, then
repeats to depth 2, etc. up to the max depth. This would be better if
we could interrupt it at any time, at which point we can play the best
move found so far. Also, we should use the results of earlier searches
to direct the deeper searches for speed. That’s a big TODO